Ticket [https://trac.osgeo.org/mapguide/ticket/253] Update to PHP 5.2.5
git-svn-id: http://svn.osgeo.org/mapguide/branches/2.0.x@2789 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Oem/php/Zend/zend_multiply.h
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
 */
 
-/* $Id: zend_multiply.h,v 1.10.2.1.2.1 2007/01/01 09:35:46 sebastian Exp $ */
+/* $Id: zend_multiply.h,v 1.10.2.1.2.2 2007/04/10 10:57:35 sniper Exp $ */
 
 #if defined(__i386__) && defined(__GNUC__)
 
@@ -35,8 +35,8 @@
 
 #define ZEND_SIGNED_MULTIPLY_LONG(a, b, lval, dval, usedval) do {		\
 	long   __lres  = (a) * (b);											\
-	double __dres  = (double)(a) * (double)(b);							\
-	double __delta = (double) __lres - __dres;							\
+	long double __dres  = (long double)(a) * (long double)(b);							\
+	long double __delta = (long double) __lres - __dres;							\
 	if ( ((usedval) = (( __dres + __delta ) != __dres))) {				\
 		(dval) = __dres;												\
 	} else {															\
